TASK1 POSTER
Mastering Layouts for Impact: I experimented with a variety of design layouts to enhance the poster's visual appeal, ensuring that every element—from images to text was thoughtfully placed to guide the viewer’s attention effectively.

Applying Filters & Merging with Them: To give the poster a cohesive and polished look, I used specific filters and incorporated the theme of Photography, blending the essence of photography into the overall design. This helped unify the campaign's message with the visual elements.

Creating for Digital Platforms: One of the most valuable aspects of this project was refining my ability to create designs that are versatile and adaptable across various digital platforms, ensuring that the poster maintains its visual appeal no matter where it's displayed.